requirements.txt
requests~=2.28.2
pytest~=7.1.2
setuptools~=66.0.0
PyYAML~=6.0
discord.py